Prelim Boruvka impl

+++ include/bucket.h
@@ -19,7 +19,7 @@ struct Bucket{
    * Checks whether the hash associated with the Bucket hashes the index to 0.
    * @return true if the index is NOT hashed to zero.
    */
-  bool contains(long index){
+  bool contains(unsigned long long int index){
     XXH64_hash_t hash = XXH64(&index, 8, bucket_seed);
     if (hash % guess_nonzero == 0)
       return true;
@@ -35,7 +35,7 @@ struct Bucket{
     //cout << "Guess: " << guess_nonzero << endl;
   }
 
-  void set_seed(long bucket_id, long sketch_seed,long random_prime){
+  void set_seed(long bucket_id, long sketch_seed, unsigned long long int random_prime){
     bucket_seed = XXH64(&bucket_id ,8, sketch_seed);
     r = bucket_seed % random_prime;
     //cout << "r : " << r << endl;
